+++ Subastas-Chivas/src/Login.css
@@ -68,3 +68,28 @@ button {
 button:hover {
     background-color: #0056b3;
 }
+
+.welcome-container {
+    text-align: center;
+    margin-top: 50px;
+  }
+  
+  .welcome-title {
+    font-size: 24px;
+    color: #333;
+  }
+  
+  .welcome-button {
+    margin-top: 20px;
+    padding: 10px 20px;
+    font-size: 16px;
+    background-color: #007bff;
+    color: white;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+  }
+  
+  .welcome-button:hover {
+    background-color: #0056b3;
+  }
